move files to components
