+++ app/containers/UIKit/VideoConferenceBlock/components/VideoConferenceBaseContainer.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { View, Text } from 'react-native';
 
-import i18n from '../../../i18n';
-import { useTheme } from '../../../theme';
-import { CustomIcon, TIconsName } from '../../CustomIcon';
+import i18n from '../../../../i18n';
+import { useTheme } from '../../../../theme';
+import { CustomIcon, TIconsName } from '../../../CustomIcon';
 import useStyle from './styles';
 
 type VideoConfMessageIconProps = {
